requirements.txt
# Function dependencies, for example:
# package>=version
wheel
google-cloud-language
gcsfs
google-auth
cloudstorage
pandas
numpy
jsons
datetime
regex
geopy